Initial version of webphp
The no framework framework for php web apps.

It's built with a lot of inspiration from konstrukt, a great framework, and gives me
a lot (defintaely not all) of the things i loved about it with *way* less code.

It's *very* crude at the moment and should be considered subject to API change
across the board. But i also believe there's something to build on here..

Let it rip!